Backend/requirements.txt
openai
flask
qdrant_client
sentence_transformers
flask_cors
firebase_admin